misc/parseq/requirements.txt
torchvision==0.20.0
torch==2.5.0
Pillow==11.0.0
PyYAML==6.0.2
pytorch_lightning==2.4.0
timm==1.0.11
nltk==3.9.1